Move legacy mod format conversion to a separate `LegacyModConverter` object
