+++ src/com/reco1l/ibancho/data/RoomMods.kt
@@ -1,6 +1,7 @@
 package com.reco1l.ibancho.data
 
 import com.reco1l.osu.multiplayer.Multiplayer
+import com.rian.osu.mods.LegacyModConverter
 import com.rian.osu.mods.ModCustomSpeed
 import com.rian.osu.mods.ModDifficultyAdjust
 import com.rian.osu.mods.ModDoubleTime
@@ -8,9 +9,8 @@ import com.rian.osu.mods.ModFlashlight
 import com.rian.osu.mods.ModHalfTime
 import com.rian.osu.mods.ModNightCore
 import com.rian.osu.utils.ModHashMap
-import com.rian.osu.utils.ModUtils
 
-class RoomMods @JvmOverloads constructor(modString: String? = null) : ModHashMap(ModUtils.convertModString(modString)) {
+class RoomMods @JvmOverloads constructor(modString: String? = null) : ModHashMap(LegacyModConverter.convertModString(modString)) {
 
     /**
      * Converts this [RoomMods] to a [String] that can be displayed to the player.
