+++ src/ru/nsu/ccfit/zuev/osu/scoring/StatisticV2.java
@@ -12,6 +12,7 @@ import com.reco1l.osu.data.ScoreInfo;
 import com.reco1l.osu.multiplayer.Multiplayer;
 import com.rian.osu.beatmap.sections.BeatmapDifficulty;
 import com.rian.osu.mods.IMigratableMod;
+import com.rian.osu.mods.LegacyModConverter;
 import com.rian.osu.mods.ModFlashlight;
 import com.rian.osu.mods.ModHidden;
 import com.rian.osu.utils.ModHashMap;
@@ -96,7 +97,7 @@ public class StatisticV2 implements Serializable {
         playerName = "";
         if (params.length < 6) return;
 
-        mod = ModUtils.convertModString(params[0]);
+        mod = LegacyModConverter.convertModString(params[0]);
         isOldScore = !params[0].contains("|");
         setForcedScore(Integer.parseInt(params[1]));
         scoreMaxCombo = Integer.parseInt(params[2]);
